Redesign block to reflect the frontend
The block previously was just a bunch of controls. This forces the block
to actually render in a similar fashion to how it looks on the frontend
while retaining the editability (so no server-side rendering in
Gutenberg)

The Gravatar Email field has been moved to the sidebar to accomodate
this.